Fix compilation errors. Still needs a few general fixes
Former-commit-id: 217cdb455d2f1ddf703d8aae95d97cba265af5f4 [formerly cfdc45031d5c3dad8cf760d2095443dcd9f11fc7]
Former-commit-id: 3c73e6981b18c186ff4962ea96338c35923b1835

+++ Robot/Src/imu_helper.cpp
@@ -81,9 +81,9 @@ void initAngularVelocityFilters(){
 }
 
 void filterAngularVelocity(IMUStruct& imu){
-    velocityFilters[VFilter::VX].update(&imu._x_Gyro, &imu._x_Gyro, 1);
-    velocityFilters[VFilter::VY].update(&imu._y_Gyro, &imu._y_Gyro, 1);
-    velocityFilters[VFilter::VZ].update(&imu._z_Gyro, &imu._z_Gyro, 1);
+    velocityFilters[static_cast<int>(VFilter::VX)].update(&imu._x_Gyro, &imu._x_Gyro, 1);
+    velocityFilters[static_cast<int>(VFilter::VY)].update(&imu._y_Gyro, &imu._y_Gyro, 1);
+    velocityFilters[static_cast<int>(VFilter::VZ)].update(&imu._z_Gyro, &imu._z_Gyro, 1);
 }
 
 } // end namespace Helpers
